Move theme toggle to InfoBar toolbar
